// src/app/core/models/User.ts
export class User {
    id: any;
    name: any;
    prenom: any;
    email: any
    username: any;
    password: any ;
    tel: any;
    address: any;
    photo: any;
    status:any;
    role!:any[];
    stateuser:any;
    fileName!: string;
    datenaissance:any;
    image!:any;
    constructor(){
   
    }
}